dev: separate into two pacakges
