+++ pkg/utils_test.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"net/url"
