+++ pkg/version.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 // Version of dbmate
 const Version = "1.2.1"
